Fix/user locator view

See merge request zacarias_link/capstone-q3!43

+++ app/services/date_ocupied_services.py
@@ -1,3 +1,6 @@
+from sqlalchemy.sql.expression import true
+from app.exc.not_permission import NotPermission
+from app.exc.not_found_error import NotFound
 from app.models.date_ocupied_model import DateOcupiedModel
 from app.services.helpers import add_in_db ,format_url_date_ocupied, check_incorrect_keys, check_missing_keys, delete_in_db
 
@@ -11,8 +14,13 @@ def get_dates_ocupied_by_paginate(data):
     page_default, per_page_default = (1, 15)
     page = data.get("page") or page_default
     per_page = data.get("per_page") or per_page_default
-    print(page, per_page)
-    date_ocupied = DateOcupiedModel.query.paginate(int(page), int(per_page),error_out=False)
+    car_id = data.get("car_id") or 0
+    unavaliable_date_id = data.get("unavaliable_date_id") or 0
+    
+    date_ocupied = DateOcupiedModel.query.filter(
+        (DateOcupiedModel.car_id==int(car_id) or true), 
+        (DateOcupiedModel.unavaliable_date_id==int(unavaliable_date_id) or True)).paginate(
+            int(page), int(per_page),error_out=False)
     
     next_url, prev_url = format_url_date_ocupied(date_ocupied, page, per_page) 
     return date_ocupied, next_url, prev_url
@@ -29,10 +37,14 @@ def post_date_ocupied_by_data(data):
     return date_ocupied
 
 
-def delete_date_ocupied_by_id(id):
+def delete_date_ocupied_by_id(id, token):
     date_to_delete = DateOcupiedModel.query.get(id)
     
     if not date_to_delete:
-        return False
+        raise NotFound
+    if date_to_delete.id != token.get("user_id"):
+        raise NotPermission
+    
     delete_in_db(date_to_delete)
+    
     return True
